+++ flutter/ios/Classes/SentryFlutterPluginApple.swift
@@ -39,6 +39,8 @@ public class SentryFlutterPluginApple: NSObject, FlutterPlugin {
         registrar.addMethodCallDelegate(instance, channel: channel)
     }
 
+    private lazy var sentryFlutter = SentryFlutter()
+
     private func registerObserver() {
         NotificationCenter.default.addObserver(self,
                                                selector: #selector(applicationDidBecomeActive),
@@ -250,7 +252,7 @@ public class SentryFlutterPluginApple: NSObject, FlutterPlugin {
         }
 
         SentrySDK.start { options in
-            self.updateOptions(arguments: arguments, options: options)
+            self.sentryFlutter.update(options: options, with: arguments)
 
             if arguments["enableAutoPerformanceTracing"] as? Bool ?? false {
                 PrivateSentrySDKOnly.appStartMeasurementHybridSDKMode = true
@@ -305,106 +307,6 @@ public class SentryFlutterPluginApple: NSObject, FlutterPlugin {
     private func closeNativeSdk(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
         SentrySDK.close()
         result("")
-    }
-
-    // swiftlint:disable:next cyclomatic_complexity
-    private func updateOptions(arguments: [String: Any], options: Options) {
-        if let dsn = arguments["dsn"] as? String {
-            options.dsn = dsn
-        }
-
-        if let isDebug = arguments["debug"] as? Bool {
-            options.debug = isDebug
-        }
-
-        if let environment = arguments["environment"] as? String {
-            options.environment = environment
-        }
-
-        if let releaseName = arguments["release"] as? String {
-            options.releaseName = releaseName
-        }
-
-        if let enableAutoSessionTracking = arguments["enableAutoSessionTracking"] as? Bool {
-            options.enableAutoSessionTracking = enableAutoSessionTracking
-        }
-
-        if let attachStacktrace = arguments["attachStacktrace"] as? Bool {
-            options.attachStacktrace = attachStacktrace
-        }
-
-        if let diagnosticLevel = arguments["diagnosticLevel"] as? String, options.debug == true {
-            options.diagnosticLevel = logLevelFrom(diagnosticLevel: diagnosticLevel)
-        }
-
-        if let sessionTrackingIntervalMillis = arguments["autoSessionTrackingIntervalMillis"] as? UInt {
-            options.sessionTrackingIntervalMillis = sessionTrackingIntervalMillis
-        }
-
-        if let dist = arguments["dist"] as? String {
-            options.dist = dist
-        }
-
-        if let enableAutoNativeBreadcrumbs = arguments["enableAutoNativeBreadcrumbs"] as? Bool {
-            options.enableAutoBreadcrumbTracking = enableAutoNativeBreadcrumbs
-        }
-
-        if let enableNativeCrashHandling = arguments["enableNativeCrashHandling"] as? Bool {
-            options.enableCrashHandler = enableNativeCrashHandling
-        }
-
-        if let maxBreadcrumbs = arguments["maxBreadcrumbs"] as? UInt {
-            options.maxBreadcrumbs = maxBreadcrumbs
-        }
-
-        if let sendDefaultPii = arguments["sendDefaultPii"] as? Bool {
-            options.sendDefaultPii = sendDefaultPii
-        }
-
-        if let maxCacheItems = arguments["maxCacheItems"] as? UInt {
-            options.maxCacheItems = maxCacheItems
-        }
-
-        if let enableWatchdogTerminationTracking = arguments["enableWatchdogTerminationTracking"] as? Bool {
-            options.enableWatchdogTerminationTracking = enableWatchdogTerminationTracking
-        }
-
-        if let sendClientReports = arguments["sendClientReports"] as? Bool {
-            options.sendClientReports = sendClientReports
-        }
-
-        if let maxAttachmentSize = arguments["maxAttachmentSize"] as? UInt {
-            options.maxAttachmentSize = maxAttachmentSize
-        }
-
-        if let captureFailedRequests = arguments["captureFailedRequests"] as? Bool {
-            options.enableCaptureFailedRequests = captureFailedRequests
-        }
-
-        if let enableAppHangTracking = arguments["enableAppHangTracking"] as? Bool {
-            options.enableAppHangTracking = enableAppHangTracking
-        }
-
-        if let appHangTimeoutIntervalMillis = arguments["appHangTimeoutIntervalMillis"] as? UInt {
-            options.appHangTimeoutInterval = TimeInterval(appHangTimeoutIntervalMillis) / 1000
-        }
-    }
-
-    private func logLevelFrom(diagnosticLevel: String) -> SentryLevel {
-        switch diagnosticLevel {
-        case "fatal":
-            return .fatal
-        case "error":
-            return .error
-        case "debug":
-            return .debug
-        case "warning":
-            return .warning
-        case "info":
-            return .info
-        default:
-            return .none
-        }
     }
 
     private func setEventOriginTag(event: Event) {